requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pinecone-client>=3.0.0
httpx>=0.25.0
sentence-transformers>=2.2.0
torch>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
